Extract shared SeasonCalendar component from the Smart Race Solver calendar
